src/main/java/com/nephest/battlenet/sc2/model/validation/NullOrNotBlankValidator.java
// Copyright (C) 2020-2023 Oleksandr Masniuk
// SPDX-License-Identifier: AGPL-3.0-or-later

package com.nephest.battlenet.sc2.model.validation;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

public class NullOrNotBlankValidator
implements ConstraintValidator<NullOrNotBlank, String>
{

    @Override
    public boolean isValid(String s, ConstraintValidatorContext constraintValidatorContext)
    {
        return s == null || !s.isBlank();
    }

}
